pyproject.toml
[project]
name = "restaurant-agent"
version = "0.1.0"
description = "Restaurant bot with triage/menu/order/reservation agents using OpenAI Agents SDK handoffs"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.8",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.1",
]
